+++ dcm2bids/dcm2bids.py
@@ -130,9 +130,8 @@ class Dcm2bids(object):
             sidecars,
             self.config["descriptions"],
             self.config.get("searchMethod", DEFAULT.searchMethod),
-            self.config.get("caseSensitive", DEFAULT.caseSensitive),
-            self.config.get("dupMethod", DEFAULT.duplicateMethod))
-
+            self.config.get("caseSensitive", DEFAULT.caseSensitive)
+        )
         parser.build_graph()
         parser.build_acquisitions(self.participant)
         parser.find_runs()
@@ -201,7 +200,7 @@ dcm2bids {}""".format(
             __version__
         ),
         epilog="""
-            Documentation at https://github.com/cbedetti/Dcm2Bids
+            Documentation at https://github.com/unfmontreal/Dcm2Bids
             """,
     )
 
@@ -262,7 +261,7 @@ dcm2bids {}""".format(
         action="store_true",
         help="""
         This option no longer exists from the script in this release.
-        See:https://github.com/cbedetti/Dcm2Bids/blob/master/README.md#defaceTpl""",
+        See:https://github.com/unfmontreal/Dcm2Bids/blob/master/README.md#defaceTpl""",
     )
 
     args = parser.parse_args()