+++ dcm2bids/utils.py
@@ -34,12 +34,7 @@ class DEFAULT(object):
     compKeys = ["SeriesNumber", "AcquisitionTime", "SidecarFilename"]
     searchMethod = "fnmatch"
     searchMethodChoices = ["fnmatch", "re"]
-
-    duplicateMethod = "run"
-    dupMethodChoices = ["run", "dup"]
     runTpl = "_run-{:02d}"
-    dupTpl = "_dup-{:02d}"
-
     caseSensitive = True
 
     # misc